[refactor] Redis subscribe 메세지 구체화 및 Message 관련 상수 관리로 필요한 값 관리
