+++ dongchimi-consumer/src/main/java/com/dcm/global/constant/MessageConstant.java
@@ -1,0 +1,7 @@
+package com.dcm.global.constant;
+
+public class MessageConstant {
+
+    public static final String SUBSCRIBE_PREFIX = "/sub/room-";
+
+}
