+++ dongchimi-consumer/src/main/java/com/dcm/message/dto/MessageRequest.java
@@ -1,0 +1,4 @@
+package com.dcm.message.dto;
+
+public record MessageRequest(Long partyId, Long chatId, String email, String message) {
+}
